requirements.txt
alembic==1.11.3
annotated-types==0.5.0
anyio==3.7.1
asyncpg==0.28.0
attrs==23.1.0
certifi==2023.7.22
click==8.1.7
envparse==0.2.0
exceptiongroup==1.1.3
fastapi==0.101.1
gunicorn==21.2.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
Mako==1.2.4
MarkupSafe==2.1.3
outcome==1.2.0
packaging==23.1
pluggy==1.2.0
psycopg2-binary==2.9.7
pydantic==2.2.1
pydantic_core==2.6.1
pytest==7.4.0
pytest-asyncio==0.21.1
python-dotenv==1.0.0
python-multipart==0.0.6
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.20
starlette==0.27.0
tomli==2.0.1
trio==0.22.2
typing_extensions==4.7.1
uvicorn==0.23.2
